[rename🚚] 도메인 분리에 따른 testCode import 수정(#21)
